requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.5
scikit-learn>=1.0.2
tqdm>=4.62.3
Pillow>=9.0.0
streamlit>=1.10.0
